requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0           # async PostgreSQL driver
alembic==1.14.0           # database migrations

# ── Vector Store ──────────────────────────────────────────────────────────────
chromadb==0.5.23

# ── LLM – Google Gemini ───────────────────────────────────────────────────────
google-generativeai==0.8.3

# ── Caching (Redis) ────────────────────────────────────────────────────────────
redis[asyncio]==5.2.0

# ── Messaging (Kafka) – production ────────────────────────────────────────────
aiokafka==0.11.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.2             # async HTTP client (for health checks / webhooks)
tenacity==9.0.0           # retry logic
orjson==3.10.11           # fast JSON serialisation

# ── Reranker ─────────────────────────────────────────────────────────────────
flashrank==0.2.9          # ultra-lite reranker (~4MB ONNX model, no Torch needed)

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2             # also used in tests via TestClient
